requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
pydantic==2.9.1
pydantic-settings==2.5.2
gspread==6.0.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
loguru==0.7.2
jinja2==3.1.6
PyYAML==6.0.3
google-api-python-client